+++ app/providers.tsx
@@ -1,4 +1,4 @@
-import { ChakraProvider } from "@chakra-ui/react";
+import { ChakraProvider, extendTheme } from "@chakra-ui/react";
 import React from "react";
 import { SessionProvider } from "../hooks/SessionContext";
 
@@ -7,8 +7,17 @@ interface props {
 }
 
 const Providers: React.FC<props> = ({ children }) => {
+  const extendedTheme = extendTheme({
+    colors: {
+      primary: {
+        base: "#1E40AF",
+        hover: "#1D4ED8",
+      },
+    },
+  });
+
   return (
-    <ChakraProvider>
+    <ChakraProvider theme={extendedTheme}>
       <SessionProvider>{children}</SessionProvider>
     </ChakraProvider>
   );
